requirements.txt
asgiref==3.6.0
check-password==0.0.2
Django==4.1.5
django-colorfield==0.8.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7
sqlparse==0.4.3
